Removed ID field from messages
The ID field is not necessary as the identity of clients can be derived from the IP address + port. The worker can store this as an immutable value instead of letting the client provide the ID. This saves bandwidth and is safer as the client cannot dictate a ID anymore.

Added enum for new message type Get which will be used by clients to retrieve the most recent overall state.

Don't send payload length to shrink message sizes even further. Those optimizations reduce the average message size by about 50%.